Refactor summary module packages
