+++ summary/src/main/kotlin/com/gdavidpb/tuindice/summary/data/repository/account/SettingsDataSource.kt
@@ -1,4 +1,4 @@
-package com.gdavidpb.tuindice.summary.data.account.source
+package com.gdavidpb.tuindice.summary.data.repository.account
 
 interface SettingsDataSource {
 	suspend fun isGetAccountOnCooldown(): Boolean
